requirements.txt
notion-client>=2.2.0
openai>=1.0.0
mcp[cli]>=1.0.0
python-dotenv>=1.0.0
